Rename pytorch_container to sagemaker_pytorch_container.
